Feature ARP-0003: Move IconWithText into defaultComponents folder
