+++ src/components/defaultComponents/iconWithText/styles/IconWithText.module.scss
@@ -1,4 +1,4 @@
-@import "../../../scssAbstracts/variables";
+@import "../../../../scssAbstracts/variables";
 
 .wrapper {
   display: flex;
